Move placeholder dataclasses + instantiate_config into dag.py; drop submit_step_to_iris
Flip the executor↔dag dependency direction: `dag.py` is now the home of
all configuration-graph data shapes (`ExecutorStep`, `InputName`,
`OutputName`, `VersionedValue`, `MirroredValue`), the deterministic
walker (`walk_config`, `upstream_steps`), and the pure substitution
passes (`instantiate_config`, `resolve_local_placeholders`). It has zero
imports from `executor.py`.

`executor.py` keeps the orchestration layer (`Executor`,
`executor_main`, `ExecutorInfo`, ...) plus the helpers that drive an
`Executor` instance at runtime (`materialize`,
`_resolve_step_output_path`). The local `from marin.execution.dag
import …` inside `collect_dependencies_and_version` is gone — that
import now lives at the top of the file. Existing call sites
(`from marin.execution.executor import ExecutorStep, …`) keep working
via top-level re-exports; the cycle is broken so re-exports are safe.

Drops `submit_step_to_iris` and its private helper
`_substitute_output_paths_only`. Both had zero production callers — the
training-step migration replaced them with `prepare_train` /
`prepare_grug_trial` + `run_train`, which bake the output_path into the
training config via `_bake_output_path_into_train_config` and submit a
plain `JobRequest`. Their tests
(`test_submit_step_to_iris_*`, `test_substitute_*`) are deleted with
them. `_resolve_step_output_path` and `resolve_local_placeholders`
remain — both are still used by the active launcher functions.
