+++ lib/marin/src/marin/execution/dag.py
@@ -1,9 +1,26 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Dependency-graph helpers for the executor.
+"""Data shapes and pure-substitution helpers for the executor pipeline.
 
-The walker emits typed events that callers consume:
+This module owns the configuration-graph primitives:
+
+  - ``ExecutorStep`` — a single step in the pipeline.
+  - ``InputName`` / ``OutputName`` / ``VersionedValue`` / ``MirroredValue`` —
+    placeholder values that appear inside a step's ``config`` and get
+    resolved at instantiation time.
+  - ``walk_config`` and ``upstream_steps`` — deterministic traversal of a
+    config graph, emitting typed events at every concrete reference.
+  - ``instantiate_config`` and ``resolve_local_placeholders`` — pure
+    substitution passes that turn placeholder configs into concrete ones.
+
+This module **does not** import from :mod:`marin.execution.executor`; the
+dependency direction is one-way (``executor → dag``). Helpers that need
+to drive an :class:`Executor` instance (``materialize``,
+``_resolve_step_output_path``) live in ``executor.py``.
+
+Walker event semantics
+----------------------
 
   - ``InputNameEvent`` — at every concrete reference to another step. Carries
     the dotted path prefix (e.g. ``"data.cache_dir"``) and the underlying
@@ -20,28 +37,261 @@ from inside `.value`). Dataclasses, dicts, lists, tuples, sets, and
 frozensets recurse element-wise; primitives are leaves.
 """
 
+import dataclasses
 import os
 from collections.abc import Iterator
 from dataclasses import dataclass, fields, is_dataclass, replace
-from typing import Any, TypeVar
+from typing import TYPE_CHECKING, Any, Generic, TypeVar
+from urllib.parse import urlparse
 
-from fray import client as fray_client
-from fray.types import Entrypoint, JobRequest, create_environment
-from rigging.filesystem import marin_prefix
-
-from marin.execution.executor import (
-    Executor,
-    ExecutorStep,
-    InputName,
-    MirroredValue,
-    OutputName,
-    VersionedValue,
-    instantiate_config,
-    output_path_of,
-)
-from marin.execution.remote import RemoteCallable, _sanitize_job_name
+from fray.types import ResourceConfig
 
 ConfigT = TypeVar("ConfigT")
+ConfigT_co = TypeVar("ConfigT_co", covariant=True)
+T_co = TypeVar("T_co", covariant=True)
+
+ExecutorFunction = Any  # Callable | None — kept loose to avoid pulling in remote types here.
+
+
+############################################################
+# Placeholder data shapes
+############################################################
+
+
+@dataclass(frozen=True)
+class ExecutorStep(Generic[ConfigT_co]):
+    """
+    An `ExecutorStep` represents a single step of a larger pipeline (e.g.,
+    transforming HTML to text).  It is specified by:
+     - a name (str), which is used to determine the `output_path`.
+     - a function `fn` (Callable), and
+     - a configuration `config` which gets passed into `fn`.
+     - a pip dependencies list (Optional[list[str]]) which are the pip dependencies required for the step.
+     These can be keys of project.optional-dependencies in the project's pyproject.toml file or any other pip package.
+
+    When a step is run, we compute the following two things for each step:
+    - `version`: represents all the upstream dependencies of the step
+    - `output_path`: the path where the output of the step are stored, based on
+    the name and a hash of the version.
+
+    The `config` is a dataclass object that recursively might have special
+    values of the following form:
+    - `InputName(step, name)`: a dependency on another `step`, resolve to the step.output_path / name
+    - `OutputName(name)`: resolves to the output_path / name
+    - `VersionedValue(value)`: a value that should be part of the version
+    The `config` is instantiated by replacing these special values with the
+    actual paths during execution.
+
+    Note: `step: ExecutorStep` is interpreted as `InputName(step, None)`.
+    """
+
+    name: str
+    fn: ExecutorFunction
+    config: ConfigT_co
+    description: str | None = None
+
+    override_output_path: str | None = None
+    """Specifies the `output_path` that should be used.  Print warning if it
+    doesn't match the automatically computed one."""
+
+    resources: ResourceConfig | None = None
+    """If set, this step is submitted as its own Fray job using these
+    resources. ``fn`` is invoked inside the submitted job.
+
+    If ``None``, behavior is determined by ``fn``: a ``RemoteCallable``
+    submits as a Fray job; a plain callable runs inline in-process.
+    """
+
+    def cd(self, name: str) -> "InputName":
+        """Refer to the `name` under `self`'s output_path."""
+        return InputName(self, name=name)
+
+    def __truediv__(self, other: str) -> "InputName":
+        """Alias for `cd`. That looks more Pythonic."""
+        return InputName(self, name=other)
+
+    def __hash__(self):
+        """Hash based on the ID (every object is different)."""
+        return hash(id(self))
+
+    def with_output_path(self, output_path: str) -> "ExecutorStep":
+        """Return a copy of the step with the given output_path."""
+        return replace(self, override_output_path=output_path)
+
+    def as_input_name(self) -> "InputName":
+        return InputName(step=self, name=None)
+
+
+@dataclass(frozen=True)
+class InputName:
+    """To be interpreted as a previous `step`'s output_path joined with `name`."""
+
+    step: ExecutorStep | None
+    name: str | None
+    block_on_step: bool = True
+    """
+    If False, the step that uses this InputName
+    will not block (or attempt to execute) `step`. We use this for
+    documenting dependencies in the config, but where that step might not have technically finished...
+
+    For instance, we sometimes use training checkpoints before the training step has finished.
+
+    These "pseudo-dependencies" still impact the hash of the step, but they don't block execution.
+    """
+
+    def cd(self, name: str) -> "InputName":
+        return InputName(self.step, name=os.path.join(self.name, name) if self.name else name)
+
+    def __truediv__(self, other: str) -> "InputName":
+        """Alias for `cd` that looks more Pythonic."""
+        return self.cd(other)
+
+    @staticmethod
+    def hardcoded(path: str) -> "InputName":
+        """
+        Sometimes we want to specify a path that is not part of the pipeline but is still relative to the prefix.
+        Try to use this sparingly.
+        """
+        return InputName(None, name=path)
+
+    def nonblocking(self) -> "InputName":
+        """
+        the step will not block on (or attempt to execute) the parent step.
+
+         (Note that if another step depends on the parent step, it will still block on it.)
+        """
+        return dataclasses.replace(self, block_on_step=False)
+
+
+def get_executor_step(run: ExecutorStep | InputName) -> ExecutorStep:
+    """
+    Helper function to extract the ExecutorStep from an InputName or ExecutorStep.
+
+    Args:
+        run (ExecutorStep | InputName): The input to extract the step from.
+
+    Returns:
+        ExecutorStep: The extracted step.
+    """
+    if isinstance(run, ExecutorStep):
+        return run
+    elif isinstance(run, InputName):
+        step = run.step
+        if step is None:
+            raise ValueError(f"Hardcoded path {run.name} is not part of the pipeline")
+        return step
+    else:
+        raise ValueError(f"Unexpected type {type(run)} for run: {run}")
+
+
+def output_path_of(step: ExecutorStep, name: str | None = None) -> InputName:
+    return InputName(step=step, name=name)
+
+
+if TYPE_CHECKING:
+
+    class OutputName(str):
+        """Type-checking stub treated as a string so defaults like THIS_OUTPUT_PATH fit `str`."""
+
+        name: str | None
+
+else:
+
+    @dataclass(frozen=True)
+    class OutputName:
+        """To be interpreted as part of this step's output_path joined with `name`."""
+
+        name: str | None
+
+
+def this_output_path(name: str | None = None):
+    return OutputName(name=name)
+
+
+# constant so we can use it in fields of dataclasses
+THIS_OUTPUT_PATH = OutputName(None)
+
+
+@dataclass(frozen=True)
+class VersionedValue(Generic[T_co]):
+    """Wraps a value, to signal that this value (part of a config) should be part of the version."""
+
+    value: T_co
+
+
+def versioned(value: T_co) -> VersionedValue[T_co]:
+    if isinstance(value, VersionedValue):
+        raise ValueError("Can't nest VersionedValue")
+    elif isinstance(value, InputName):
+        # TODO: We have also run into Versioned([InputName(...), ...])
+        raise ValueError("Can't version an InputName")
+
+    return VersionedValue(value)
+
+
+def ensure_versioned(value: VersionedValue[T_co] | T_co) -> VersionedValue[T_co]:
+    """
+    Ensure that the value is wrapped in a VersionedValue. If it is already wrapped, return it as is.
+    """
+    return value if isinstance(value, VersionedValue) else VersionedValue(value)
+
+
+def unwrap_versioned_value(value: VersionedValue[T_co] | T_co) -> T_co:
+    """
+    Unwrap the value if it is a VersionedValue, otherwise return the value as is.
+
+    Recurses into dataclasses, dicts and lists to unwrap any nested VersionedValue instances.
+    This method cannot handle InputName, OutputName, or ExecutorStep instances inside VersionedValue as
+    their values depend on execution results.
+    """
+
+    def recurse(obj: Any):
+        if isinstance(obj, MirroredValue):
+            return recurse(obj.value)
+        if isinstance(obj, VersionedValue):
+            return recurse(obj.value)
+        if isinstance(obj, OutputName | InputName | ExecutorStep):
+            raise ValueError(f"Cannot unwrap VersionedValue containing {type(obj)}: {obj}")
+        if is_dataclass(obj):
+            result = {}
+            for field in fields(obj):
+                val = getattr(obj, field.name)
+                result[field.name] = recurse(val)
+            return replace(obj, **result)
+        if isinstance(obj, dict):
+            return {k: recurse(v) for k, v in obj.items()}
+        if isinstance(obj, list):
+            return [recurse(x) for x in obj]
+        return obj
+
+    return recurse(value)  # type: ignore
+
+
+@dataclass(frozen=True)
+class MirroredValue(Generic[T_co]):
+    """Wraps a path value to signal that it should be mirrored from any marin regional bucket.
+
+    At config instantiation time, the path is resolved to the local marin prefix.
+    Before step execution, the executor copies the data from whichever region has it.
+    """
+
+    value: T_co
+    budget_gb: float = 10
+
+
+def mirrored(value: str | VersionedValue[str], budget_gb: float = 10) -> MirroredValue:
+    """Mark a path for cross-region mirroring with a transfer budget.
+
+    Usage: input_path=mirrored(versioned("documents/stackexchange/..."), budget_gb=50)
+    """
+    if isinstance(value, MirroredValue):
+        raise ValueError("Can't nest MirroredValue")
+    return MirroredValue(value=value, budget_gb=budget_gb)
+
+
+############################################################
+# Walker
+############################################################
 
 
 @dataclass(frozen=True)
@@ -174,137 +424,74 @@ def upstream_steps(obj: Any) -> list[ExecutorStep]:
     return result
 
 
-def materialize(
-    config: ConfigT,
-    *,
-    prefix: str | None = None,
-    output_path: str | None = None,
-) -> ConfigT:
-    """Run any `ExecutorStep`s embedded in `config`, then return a copy of
-    `config` with all placeholder paths substituted.
+############################################################
+# Substitution
+############################################################
 
-    Composes three pieces:
 
-      1. ``upstream_steps(config)`` — find embedded ``ExecutorStep``s.
-      2. ``Executor(prefix=...).run(steps)`` — submit them as sub-jobs and
-         block on completion. The distributed ``step_lock`` keeps concurrent
-         callers (sweep members, multi-VM TPU tasks) coordinated; only one
-         actually runs each step, the rest read ``STATUS_SUCCESS`` and skip.
-      3. ``instantiate_config(config, output_path=<resolved>,
-         output_paths=executor.output_paths, prefix=prefix)`` — substitute
-         ``InputName`` / ``OutputName`` / ``VersionedValue`` / ``ExecutorStep``
-         placeholders using the just-computed paths.
+def _is_relative_path(url_or_path: str) -> bool:
+    parsed_url = urlparse(url_or_path)
+    if parsed_url.scheme:
+        return False
+    return not url_or_path.startswith("/")
 
-    Args:
-        config: A launcher config dataclass that may embed ``ExecutorStep``s
-            and placeholder values.
-        prefix: Storage prefix for newly-submitted sub-jobs. Defaults to
-            ``marin_prefix()`` (the worker's regional ``gs://marin-{R}``
-            bucket), so upstream data is co-located with training.
-        output_path: Concrete output path for the current step, used to
-            resolve ``OutputName(name=...)`` placeholders inside ``config``.
-            If ``None``, ``materialize`` reads ``config.output_path``. For
-            callers whose config type does not expose ``output_path``, pass
-            it explicitly.
 
-    Returns:
-        A copy of ``config`` with all placeholders substituted to concrete
-        paths. A config containing no placeholders round-trips unchanged
-        (idempotent — no sub-jobs submitted).
+def _make_prefix_absolute_path(prefix: str, override_path: str) -> str:
+    if _is_relative_path(override_path):
+        override_path = os.path.join(prefix, override_path)
+    return override_path
+
+
+def instantiate_config(config: Any, output_path: str, output_paths: dict[ExecutorStep, str], prefix: str) -> Any:
     """
-    resolved_prefix = prefix if prefix is not None else marin_prefix()
-
-    steps = upstream_steps(config)
-
-    # Idempotence guard: if no sub-steps reference the config, skip the
-    # `Executor.run` path entirely. `Executor.run([])` would otherwise still
-    # write out an executor-info JSON to GCS (`write_infos` runs unconditionally
-    # in `executor.run`), which is both pointless and an unwanted I/O side
-    # effect for a placeholder-free config.
-    if steps:
-        executor_info_base_path = os.path.join(resolved_prefix, "experiments")
-        executor = Executor(
-            prefix=resolved_prefix,
-            executor_info_base_path=executor_info_base_path,
-        )
-        output_paths: dict[ExecutorStep, str] = executor.run(steps=steps)
-    else:
-        output_paths = {}
-
-    if output_path is None:
-        current_output_path = config.output_path  # type: ignore[attr-defined]
-    else:
-        current_output_path = output_path
-
-    if isinstance(current_output_path, OutputName):
-        raise TypeError(
-            "materialize(config): output_path is still an OutputName "
-            "placeholder. The launcher / job-submission layer must resolve the "
-            "current step's output_path to a concrete string before calling "
-            "the worker function. Got: "
-            f"{current_output_path!r}"
-        )
-
-    return instantiate_config(
-        config=config,
-        output_path=current_output_path,
-        output_paths=output_paths,
-        prefix=resolved_prefix,
-    )
-
-
-def _resolve_step_output_path(step: ExecutorStep, *, prefix: str | None = None) -> str:
-    """Compute the concrete output path for ``step`` without running anything.
-
-    Drives ``Executor.compute_version`` (which only walks the dependency graph
-    and computes hashes — no GCS I/O, no job submission) far enough to populate
-    ``Executor.output_paths[step]``.
-    """
-    resolved_prefix = prefix if prefix is not None else marin_prefix()
-    executor_info_base_path = os.path.join(resolved_prefix, "experiments")
-    executor = Executor(
-        prefix=resolved_prefix,
-        executor_info_base_path=executor_info_base_path,
-    )
-    executor.compute_version(step, is_pseudo_dep=False)
-    return executor.output_paths[step]
-
-
-def _substitute_output_paths_only(config: ConfigT, output_path: str) -> ConfigT:
-    """Replace ``OutputName`` placeholders in ``config`` with paths under
-    ``output_path``. Other placeholders (``InputName``, ``ExecutorStep``,
-    ``VersionedValue``, ``MirroredValue``) are left intact for the worker's
-    ``materialize`` call to resolve in the worker's own region.
-
-    Recurses *into* ``MirroredValue`` and ``VersionedValue`` to surface any
-    nested ``OutputName``, but rebuilds the wrapper rather than resolving it —
-    the worker still needs to see the wrapper to apply its region-aware
-    semantics.
+    Return a "real" config where all the special values (e.g., `InputName`,
+    `OutputName`, and `VersionedValue`) have been replaced with
+    the actual paths that they represent.
+    `output_path`: represents the output path of the current step.
+    `output_paths`: a dict from `ExecutorStep` to their output paths.
     """
 
-    def join_path(name: str | None) -> str:
+    def join_path(output_path: str, name: str | None) -> str:
         return os.path.join(output_path, name) if name else output_path
 
-    def recurse(obj: Any) -> Any:
+    def recurse(obj: Any):
         if obj is None:
             return None
-        if isinstance(obj, OutputName):
-            return join_path(obj.name)
+
+        if isinstance(obj, ExecutorStep):
+            obj = output_path_of(obj)
+
         if isinstance(obj, MirroredValue):
-            return replace(obj, value=recurse(obj.value))
-        if isinstance(obj, VersionedValue):
-            return replace(obj, value=recurse(obj.value))
-        # InputName / ExecutorStep are deferred to the worker's materialize().
-        if isinstance(obj, (InputName, ExecutorStep)):
-            return obj
-        if is_dataclass(obj) and not isinstance(obj, type):
-            updates = {field.name: recurse(getattr(obj, field.name)) for field in fields(obj)}
-            return replace(obj, **updates)
-        if isinstance(obj, list):
+            inner = recurse(obj.value)
+            # Resolve to mirror:// protocol — MirrorFileSystem handles cross-region copying
+            if isinstance(inner, str) and not inner.startswith("mirror://"):
+                return f"mirror://{inner}"
+            return inner
+
+        if isinstance(obj, InputName):
+            if obj.step is None:
+                return _make_prefix_absolute_path(prefix, obj.name)
+            else:
+                return join_path(output_paths[obj.step], obj.name)
+        elif isinstance(obj, OutputName):
+            return join_path(output_path, obj.name)
+        elif isinstance(obj, VersionedValue):
+            return obj.value
+        elif is_dataclass(obj):
+            # Recurse through dataclasses
+            result = {}
+            for field in fields(obj):
+                value = getattr(obj, field.name)
+                result[field.name] = recurse(value)
+            return replace(obj, **result)
+        elif isinstance(obj, list):
+            # Recurse through lists
             return [recurse(x) for x in obj]
-        if isinstance(obj, dict):
-            return {k: recurse(v) for k, v in obj.items()}
-        return obj
+        elif isinstance(obj, dict):
+            # Recurse through dicts
+            return dict((i, recurse(x)) for i, x in obj.items())
+        else:
+            return obj
 
     return recurse(config)
 
@@ -335,8 +522,9 @@ def resolve_local_placeholders(config: ConfigT, output_path: str) -> ConfigT:
         if isinstance(obj, MirroredValue):
             return replace(obj, value=recurse(obj.value))
         if isinstance(obj, VersionedValue):
-            # Version tracking only matters for hash assembly, which was done
-            # by `_resolve_step_output_path`; unwrap fully here.
+            # Version tracking only matters for hash assembly, which the
+            # caller has already performed via the executor's version pass;
+            # unwrap fully here.
             return recurse(obj.value)
         if isinstance(obj, (InputName, ExecutorStep)):
             return obj
@@ -350,44 +538,3 @@ def resolve_local_placeholders(config: ConfigT, output_path: str) -> ConfigT:
         return obj
 
     return recurse(config)
-
-
-def submit_step_to_iris(step: ExecutorStep) -> None:
-    """Submit ``step`` as a top-level Iris job and block until it terminates.
-
-    The submitted job runs ``step.fn(resolved_config)`` on a worker chosen by
-    ``step.resources``. ``resolved_config`` is ``step.config`` with the
-    current step's ``OutputName`` placeholders pre-substituted to a concrete
-    path; upstream ``InputName(step=…)`` / ``ExecutorStep`` references are
-    left for the worker to resolve via ``materialize`` in its own region.
-
-    Args:
-        step: The ExecutorStep to submit. Must have ``resources`` set.
-
-    Raises:
-        ValueError: if ``step.resources`` is ``None``.
-        Any exception raised by the submitted job.
-    """
-    if step.resources is None:
-        raise ValueError(
-            f"submit_step_to_iris requires step.resources to be set on {step.name!r}; "
-            f"got None. Set ExecutorStep(resources=ResourceConfig.with_tpu(...))."
-        )
-
-    step_output_path = _resolve_step_output_path(step)
-    resolved_config = _substitute_output_paths_only(step.config, step_output_path)
-
-    # Unwrap RemoteCallable so the inner fn runs directly in the submitted
-    # job; re-submitting from inside would double-launch.
-    inner_fn = step.fn.fn if isinstance(step.fn, RemoteCallable) else step.fn
-
-    job_request = JobRequest(
-        name=_sanitize_job_name(step.name),
-        entrypoint=Entrypoint.from_callable(inner_fn, args=[resolved_config]),
-        resources=step.resources,
-        environment=create_environment(),
-    )
-
-    client = fray_client.current_client()
-    handle = client.submit(job_request)
-    handle.wait(raise_on_failure=True)
